Fix ubuntu image
The image was previously in a WIP state where I was trying to get
the image size down. This is not a stripped image and is still
1.4GB or so.

The start.sh script now takes the port number to easily make multiple
containers on different ports for testing purposes. I added in the
settings file stuff from the alpine image into the start script
but I ended up commenting it out because I think this image works
without it. Settings have persisted over multiple containers.

The image has been renamed "zedchance/theia-sierra-ubuntu".