Add stop_idles.sh script
This script starts by looping over all of the running docker containers,
it checks the log for a line indicating that the theia ide window
has been closed. It then checks the timestamp to see if that log
entry is older than 30 minutes, and if so it stops that container.